+++ CHANGELOG.md
@@ -17,8 +17,9 @@
 
 # 1.3.36
 
-* [X] #139 修复某些情况下无法正常渲染与预览DDS图片的问题 / Fixed a problem that DDS images could not be properly rendered and previewed in some situations
 * [X] 完善Stellaris 4.0的CWT规则文件 / Optimize CWT config files for Stellaris 4.0
+* [X] #139 修复某些情况下无法正常渲染与预览DDS图片的问题 / Fixed a problem that DDS images could not be properly rendered and previewed in some situations
+* [X] 可以分别配置是否启用内置、全局本地、项目本地的规则分组 / Allow to configure whether to enable built-in, global-local, and project-local config groups separately
 
 ## 1.3.36